Add a 'shared secret' conf for cloak key
The key shouldn't have been included in the config in plaintext in the
first place, whoops!

The impact of the key being present is not that it reveals hostmasks
since the cloaking process is lossy, but it does make it possible to
narrow down a hostmask to a much smaller number of possible inputs (even
if it can't be definitively identified as one of them).